LayoutTests/imported/w3c/web-platform-tests/svg/animations/seeking-events-8-expected.txt


FAIL Seeking backwards to a new interval dispatches both 'endEvent' and 'beginEvent' assert_equals: Expected endEvent event, but got beginEvent event instead expected "endEvent" but got "beginEvent"
